# backend/users-api/docs/openapi/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API that handles the users information.",
        "title": "Users API",
        "termsOfService": "https://www.linkedin.com/in/emilianokohmann/",
        "contact": {
            "name": "Emiliano Kohmann",
            "url": "https://www.linkedin.com/in/emilianokohmann/",
            "email": "emikohmann@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/users": {
            "get": {
                "description": "Return the users information fetching information from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Return a list of users.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "List offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.ListUsersResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    }
                }
            },
            "post": {
                "description": "Store the user information against the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Store the user information.",
                "parameters": [
                    {
                        "description": "User to save",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SaveUserRequestHTTP"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.SaveUserResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    }
                }
            }
        },
        "/users/{userID}": {
            "get": {
                "description": "Return the user information fetching information from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Return the user information.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user to get",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.GetUserResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the user information against the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Updates the user information.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user to get",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateUserRequestHTTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.UpdateUserResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the user information against the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete the user information.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user to delete",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.DeleteUserResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorHTTP"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.APIErrorHTTP": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Some information not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "http.DeleteUserResponseHTTP": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "http.GetUserResponseHTTP": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "description": {
                    "type": "string",
                    "example": "The iPhone 13 display has rounded corners"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://www.macstation.com.ar/img/productos/2599-2.jpg"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_updated": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 13 128GB 4GB RAM"
                },
                "price": {
                    "type": "number",
                    "example": 729.99
                },
                "restrictions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://contactcenter.macstation.com.ar/web/image?unique=ed3cc51"
                }
            }
        },
        "http.ListUsersResponseHTTP": {
            "type": "object",
            "properties": {
                "paging": {
                    "$ref": "#/definitions/http.PagingResponseHTTP"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.UserResponseHTTP"
                    }
                }
            }
        },
        "http.PagingResponseHTTP": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "offset": {
                    "type": "integer",
                    "example": 50
                },
                "total": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "http.SaveUserRequestHTTP": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "The iPhone 13 display has rounded corners"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://www.macstation.com.ar/img/productos/2599-2.jpg"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 13 128GB 4GB RAM"
                },
                "price": {
                    "type": "number",
                    "example": 729.99
                },
                "restrictions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://contactcenter.macstation.com.ar/web/image?unique=ed3cc51"
                }
            }
        },
        "http.SaveUserResponseHTTP": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "description": {
                    "type": "string",
                    "example": "The iPhone 13 display has rounded corners"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://www.macstation.com.ar/img/productos/2599-2.jpg"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_updated": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 13 128GB 4GB RAM"
                },
                "price": {
                    "type": "number",
                    "example": 729.99
                },
                "restrictions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://contactcenter.macstation.com.ar/web/image?unique=ed3cc51"
                }
            }
        },
        "http.UpdateUserRequestHTTP": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "The iPhone 13 display has rounded corners"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://www.macstation.com.ar/img/productos/2599-2.jpg"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 13 128GB 4GB RAM"
                },
                "price": {
                    "type": "number",
                    "example": 729.99
                },
                "restrictions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://contactcenter.macstation.com.ar/web/image?unique=ed3cc51"
                }
            }
        },
        "http.UpdateUserResponseHTTP": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "description": {
                    "type": "string",
                    "example": "The iPhone 13 display has rounded corners"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://www.macstation.com.ar/img/productos/2599-2.jpg"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_updated": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 13 128GB 4GB RAM"
                },
                "price": {
                    "type": "number",
                    "example": 729.99
                },
                "restrictions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://contactcenter.macstation.com.ar/web/image?unique=ed3cc51"
                }
            }
        },
        "http.UserResponseHTTP": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "description": {
                    "type": "string",
                    "example": "The iPhone 13 display has rounded corners"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://www.macstation.com.ar/img/productos/2599-2.jpg"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "last_updated": {
                    "type": "string",
                    "example": "2023-02-23T21:46:28.366Z"
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 13 128GB 4GB RAM"
                },
                "price": {
                    "type": "number",
                    "example": 729.99
                },
                "restrictions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://contactcenter.macstation.com.ar/web/image?unique=ed3cc51"
                }
            }
        }
    }
}